requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
sqlite3==3.39.4
datetime==4.9.0
logging==0.5.1.2
